+++ src/renderer/src/extensions/download_management/DownloadManager.ts
@@ -57,43 +57,8 @@ const MAX_STALL_RESETS = 2;
 
 export type RedownloadMode = "always" | "never" | "ask" | "replace";
 
-export class AlreadyDownloaded extends Error {
-  private mFileName: string;
-  private mId: string;
-  constructor(fileName: string, id?: string) {
-    super("File already downloaded");
-    Error.captureStackTrace(this, this.constructor);
-
-    this.name = this.constructor.name;
-    this.mFileName = fileName;
-    this.mId = id;
-  }
-
-  public get fileName(): string {
-    return this.mFileName;
-  }
-
-  public get downloadId() {
-    return this.mId;
-  }
-
-  public set downloadId(id: string) {
-    this.mId = id;
-  }
-}
-
-export class DownloadIsHTML extends Error {
-  private mUrl: string;
-  constructor(inputUrl: string) {
-    super("");
-    this.name = this.constructor.name;
-    this.mUrl = inputUrl;
-  }
-
-  public get url() {
-    return this.mUrl;
-  }
-}
+import { AlreadyDownloaded, DownloadIsHTML } from "@vortex/shared/errors";
+export { AlreadyDownloaded, DownloadIsHTML };
 
 function isHTMLHeader(headers: http.IncomingHttpHeaders) {
   return (
